Move Client struct to a core sub-module
